+++ jdaviz/configs/imviz/tests/test_viewers.py
@@ -1,6 +1,7 @@
 import numpy as np
 import pytest
 from astropy import units as u
+from astropy.coordinates import SkyCoord
 from astropy.nddata import NDData
 from regions import CirclePixelRegion, PixCoord
 
@@ -8,7 +9,11 @@ from jdaviz.app import Application
 from jdaviz.core.config import get_configuration
 from jdaviz.configs.imviz.helper import Imviz
 from jdaviz.configs.imviz.plugins.viewers import ImvizImageView
-from jdaviz.configs.imviz.tests.utils import BaseImviz_WCS_NoWCS
+from jdaviz.configs.imviz.tests.utils import (
+    BaseImviz_WCS_NoWCS, create_example_gwcs
+)
+
+from numpy.testing import assert_allclose
 
 
 @pytest.mark.parametrize(
@@ -202,6 +207,97 @@ def test_catalog_in_image_viewer(imviz_helper, image_2d_wcs, source_catalog):
     dm.layer.selected = ['my_catalog']
     dm.remove_from_app()
     assert 'my_catalog' not in imviz_helper.app.data_collection.labels
+
+
+def test_get_viewport_sky_region_wcs(imviz_helper, image_hdu_wcs):
+    imviz_helper.load(image_hdu_wcs)
+    viewer = imviz_helper.viewers['imviz-0']
+    region = viewer.get_viewport_region()
+
+    expected_vertices = SkyCoord(
+        [337.52124634, 337.52124633, 337.51664042, 337.51664038],
+        [-20.83297927, -20.83118685, -20.83118681, -20.83297923],
+        unit=u.deg
+    )
+
+    assert_allclose(
+        region.vertices.separation(expected_vertices).arcsec,
+        0, atol=5
+    )
+
+
+def test_get_viewport_sky_region_gwcs(imviz_helper):
+    shape = (10, 10)
+    data = np.ones(shape)
+    ndd = NDData(data=data, wcs=create_example_gwcs(shape))
+
+    imviz_helper.load(ndd)
+    viewer = imviz_helper.viewers['imviz-0']
+    region = viewer.get_viewport_region()
+
+    expected_vertices = SkyCoord(
+        [197.89244966, 197.89271435, 197.89256153, 197.89229684],
+        [-1.36574339, -1.36559061, -1.36532599, -1.36547877],
+        unit=u.deg
+    )
+
+    assert_allclose(
+        region.vertices.separation(expected_vertices).arcsec,
+        0, atol=5
+    )
+
+
+def test_get_viewport_sky_no_wcs(imviz_helper):
+    shape = (10, 10)
+    data = np.ones(shape)
+    ndd = NDData(data=data)
+
+    imviz_helper.load(ndd)
+    viewer = imviz_helper.viewers['imviz-0']
+
+    with pytest.warns(UserWarning, match="does not have valid WCS"):
+        viewer.get_viewport_region()
+
+
+def test_get_viewport_pixel_region(imviz_helper):
+    shape = (10, 10)
+    data = np.ones(shape)
+    ndd = NDData(data=data, wcs=create_example_gwcs(shape))
+
+    imviz_helper.load(ndd)
+    viewer = imviz_helper.viewers['imviz-0']
+
+    data_label = imviz_helper.app.data_collection[0].label
+    region = viewer.get_viewport_region('pixel', data_label)
+    assert_allclose(
+        region.vertices.x, [-0.5, -0.5, 9.5, 9.5]
+    )
+    assert_allclose(
+        region.vertices.y, [-0.5, 9.5, 9.5, -0.5]
+    )
+
+
+def test_get_viewport_pixel_region_bad_label(imviz_helper):
+    shape = (10, 10)
+    data = np.ones(shape)
+    ndd = NDData(data=data)
+    imviz_helper.load(ndd, data_label='label 1')
+
+    with pytest.raises(ValueError, match="No data found with label xyz"):
+        viewer = imviz_helper.viewers['imviz-0']
+        viewer.get_viewport_region('pixel', 'xyz')
+
+
+def test_get_viewport_pixel_region_no_label(imviz_helper):
+    shape = (10, 10)
+    data = np.ones(shape)
+    ndd = NDData(data=data)
+    imviz_helper.load(ndd, data_label='label 1')
+
+    with pytest.raises(ValueError,
+                       match="`get_viewport_region` requires a data label"):
+        viewer = imviz_helper.viewers['imviz-0']
+        viewer.get_viewport_region('pixel')
 
 
 class TestDeleteData(BaseImviz_WCS_NoWCS):